changes to previous implementation
